Renames the r-engine module into r-script-engine.
