+++ pom.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
         <modelVersion>4.0.0</modelVersion>
-        <artifactId>orbisgis-view-plugins</artifactId>
+        <artifactId>orbisgis-official-plugins</artifactId>
         <packaging>pom</packaging>
         <version>5.1</version>
-        <name>OrbisGIS View Plugins</name>
+        <name>OrbisGIS Official Plugins</name>
         <parent>
                 <groupId>org.orbisgis</groupId>
                 <artifactId>orbisgis-nexus</artifactId>
@@ -19,7 +19,7 @@
                 <module>memory</module>
                 <module>groovyConsole</module>
                 <module>mailto</module>
-                <module>r-engine</module>
+                <module>r-script-engine</module>
                 <module>r-console</module>
                 <module>chartView</module>
         </modules>
