+++ r-console/pom.xml
@@ -104,7 +104,7 @@
         </dependency>
         <dependency>
             <groupId>org.orbisgis</groupId>
-            <artifactId>r_engine</artifactId>
+            <artifactId>r-script-engine</artifactId>
             <version>5.1.0</version>
         </dependency>
     </dependencies>
