+++ r-script-engine/pom.xml
@@ -1,8 +1,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>r_engine</artifactId>
-    <name>R engine</name>
+    <artifactId>r-script-engine</artifactId>
+    <name>R script engine</name>
     <version>5.1.0</version>
     <url>http://orbisgis.org</url>
     <parent>
@@ -11,7 +11,7 @@
         <version>5.1</version>
         <relativePath>../</relativePath>
     </parent>
-    <description>A R engine based on Renjin.</description>
+    <description>A R script engine based on Renjin.</description>
     <packaging>bundle</packaging>
     <build>
         <plugins>
