+++ r-script-engine/src/main/resources/org/orbisgis/rscriptengine/icon.png
